genios_back/requirements.txt
django-rest-framework
django-cors-headers
requests
pandas
load_dotenv